Move transaction to types crate
